chore: replace socketio engine with v3
