+++ test/core/scripts/hello_socketio.json
@@ -1,49 +1,159 @@
 {
   "config": {
-      "target": "http://127.0.0.1:9091",
-      "tls": {
-        "rejectUnauthorized": true
-      },
-      "phases": [
-        {"duration": 1, "arrivalRate": 1}
-      ]
+    "target": "http://127.0.0.1:9091",
+    "tls": {
+      "rejectUnauthorized": true
+    },
+    "phases": [{ "duration": 1, "arrivalRate": 1 }]
   },
   "scenarios": [
     {
       "engine": "socketio",
       "flow": [
-        {"emit": { "channel": "echo", "data": "hello", "response": { "channel": "echoed", "data": "hello"} }},
-        {"emit": { "channel": "echo", "data": "world", "response": { "channel": "echoed", "data": "world"} }},
-        {"think": 1},
-        {"emit": { "channel": "echo", "data": "do not care about the response" }},
-        {"emit": { "channel": "echo", "data": "still do not care about the response" }},
-        {"think": 1},
-        {"emit": { "channel": "echo", "data": "{\"key\":\"value\"}", "response": { "channel": "echoed", "data": "{\"key\":\"value\"}"} }},
-        {"think": 1},
-        {"emit": { "channel": "echo", "data": "random number: {{ $randomNumber(1, 20) }}" }},
-        {"emit": { "channel": "echo", "data": {"key": "{{ $randomString(10) }}"}, "response": { "channel": "echoed", "capture": {"json": "$.key", "as": "rand" } }}},
-        {"think": 1},
-        {"emit": { "channel": "echo", "data": {"key": "{{ rand }}"}, "response": { "channel": "echoed", "match": {"json": "$.key", "value": "{{ rand }}" } }}},
-        {"think": 1},
-        {"emit": { "channel": "echo", "data": "ping", "acknowledge": {"match": {"value": "pong"}} }},
-        {"think": 1},
-        {"emit": { "channel": "echo", "data": "ping", "acknowledge": {"match": {"json": "$.1.answer", "value": 42}} }},
-        {"think": 1},
+        {
+          "emit": {
+            "channel": "echo",
+            "data": "hello"
+          },
+          "response": { "channel": "echoed", "data": "hello" }
+        },
+        {
+          "emit": {
+            "channel": "echo",
+            "data": "world"
+          },
+          "response": { "channel": "echoed", "data": "world" }
+        },
+        { "think": 1 },
+        {
+          "emit": {
+            "channel": "echo",
+            "data": "do not care about the response"
+          }
+        },
+        {
+          "emit": {
+            "channel": "echo",
+            "data": "still do not care about the response"
+          }
+        },
+        { "think": 1 },
+        {
+          "emit": {
+            "channel": "echo",
+            "data": "{\"key\":\"value\"}"
+          },
+          "response": { "channel": "echoed", "data": "{\"key\":\"value\"}" }
+        },
+        { "think": 1 },
+        {
+          "emit": {
+            "channel": "echo",
+            "data": "random number: {{ $randomNumber(1, 20) }}"
+          }
+        },
+        {
+          "emit": {
+            "channel": "echo",
+            "data": { "key": "{{ $randomString(10) }}" }
+          },
+          "response": {
+            "channel": "echoed",
+            "capture": { "json": "$.key", "as": "rand" }
+          }
+        },
+        { "think": 1 },
+        {
+          "emit": {
+            "channel": "echo",
+            "data": { "key": "{{ rand }}" }
+          },
+          "response": {
+            "channel": "echoed",
+            "match": { "json": "$.key", "value": "{{ rand }}" }
+          }
+        },
+        { "think": 1 },
+        {
+          "emit": {
+            "channel": "echo",
+            "data": "ping"
+          },
+          "acknowledge": { "match": { "value": "pong" } }
+        },
+        { "think": 1 },
+        {
+          "emit": {
+            "channel": "echo",
+            "data": "ping"
+          },
+          "acknowledge": { "match": { "json": "$.1.answer", "value": 42 } }
+        },
+        { "think": 1 },
 
-        {"loop": [
-          {"emit": { "channel": "echo", "data": "outer {{ outerElement }} {{ outerCount }}" }},
-          {"emit": { "channel": "echo", "data": "count:inc", "acknowledge": {"match": {"value": "count"}}}},
-          {"think": 1},
-          {"emit": { "channel": "echo", "data": "outer - element: {{ outerElement }} count: {{ outerCount }}" }},
-          {"think": 1},
-          {"loop": [
-            {"emit": { "channel": "echo", "data": "count:inc", "acknowledge": {"match": {"value": "count"}}}},
-            {"emit": { "channel": "echo", "data": "inner - element: {{ innerElement }} count: {{ innerCount }}" }}
-          ], "loopElement": "innerElement", "loopValue": "innerCount", "over": ["one", "two", "three"]}
-        ], "loopElement": "outerElement", "loopValue":"outerCount", "count": 3},
+        {
+          "loop": [
+            {
+              "emit": {
+                "channel": "echo",
+                "data": "outer {{ outerElement }} {{ outerCount }}"
+              }
+            },
+            {
+              "emit": {
+                "channel": "echo",
+                "data": "count:inc"
+              },
+              "acknowledge": { "match": { "value": "count" } }
+            },
+            { "think": 1 },
+            {
+              "emit": {
+                "channel": "echo",
+                "data": "outer - element: {{ outerElement }} count: {{ outerCount }}"
+              }
+            },
+            { "think": 1 },
+            {
+              "loop": [
+                {
+                  "emit": {
+                    "channel": "echo",
+                    "data": "count:inc"
+                  },
+                  "acknowledge": { "match": { "value": "count" } }
+                },
+                {
+                  "emit": {
+                    "channel": "echo",
+                    "data": "inner - element: {{ innerElement }} count: {{ innerCount }}"
+                  }
+                }
+              ],
+              "loopElement": "innerElement",
+              "loopValue": "innerCount",
+              "over": ["one", "two", "three"]
+            }
+          ],
+          "loopElement": "outerElement",
+          "loopValue": "outerCount",
+          "count": 3
+        },
 
-        {"emit": {"channel": "echo", "data": "count:read", "acknowledge": {"match": {"json": "$.1.answer", "value": 12}}}},
-        {"emit": {"channel": "echo", "data": "count:reset", "acknowledge": {"match": {"json": "$.1.answer", "value": 0}}}}
+        {
+          "emit": {
+            "channel": "echo",
+            "data": "count:read"
+          },
+          "acknowledge": { "match": { "json": "$.1.answer", "value": 12 } }
+        },
+        {
+          "emit": {
+            "channel": "echo",
+            "data": "count:reset"
+          },
+          "acknowledge": { "match": { "json": "$.1.answer", "value": 0 } }
+        }
       ]
     }
   ]
